requirement.txt
langchain
langgraph
python-dotenv
langchain-google-genai
streamlit
langchain-community
langserve
fastapi
uvicorn
ipykernel
bs4
pypdf
faiss-cpu
chromadb
wikipedia
arxiv
langchainhub
langchain-groq
cassio
beautifulsoup4